Moved GameManager and BusPlayerAdapter into impl package
