+++ src/test/java/de/dhbw/swe/betterjeweled/impl/BusPlayerAdapterTest.java
@@ -1,6 +1,7 @@
-package de.dhbw.swe.betterjeweled.core;
+package de.dhbw.swe.betterjeweled.impl;
 
 import com.google.common.eventbus.*;
+import de.dhbw.swe.betterjeweled.core.*;
 import org.junit.jupiter.api.*;
 import org.mockito.*;
